remove mito transport model. mito now serves as a demand generator, the actual assignment is done within silo using the same infrastructure as the direct coupling. This removes duplicated code. The matsim transport model now requires a matsim scenario assembler (implemented for mito and for the simple direct coupling).
This approach allows us:
- to maintain code better as we do not have to maintain mito and matsim transport model (i.e. for travel time updates)
- pass a consistent config to the whole silo mito matsim chain to be able to configure the assignment through the config
- add matsim contribs like noise or emissions from within silo and use the respective output data in silo
