+++ useCases/maryland/src/main/java/de/tum/bgu/msm/run/ModelBuilderMstm.java
@@ -1,20 +1,14 @@
-package de.tum.bgu.msm;
+package de.tum.bgu.msm.run;
 
+import de.tum.bgu.msm.container.DataContainer;
 import de.tum.bgu.msm.container.ModelContainer;
-import de.tum.bgu.msm.models.relocation.moves.RegionProbabilityStrategyImpl;
-import de.tum.bgu.msm.schools.DataContainerWithSchools;
 import de.tum.bgu.msm.data.dwelling.DwellingFactory;
 import de.tum.bgu.msm.data.household.HouseholdFactory;
-import de.tum.bgu.msm.data.mito.MitoDataConverterMuc;
 import de.tum.bgu.msm.data.person.PersonFactory;
 import de.tum.bgu.msm.matsim.MatsimTransportModel;
+import de.tum.bgu.msm.matsim.SimpleMatsimScenarioAssembler;
 import de.tum.bgu.msm.matsim.ZoneConnectorManager;
-import de.tum.bgu.msm.mito.MitoTransportModel;
-import de.tum.bgu.msm.models.EducationModelMuc;
-import de.tum.bgu.msm.models.MarriageModelMuc;
-import de.tum.bgu.msm.models.autoOwnership.CreateCarOwnershipModel;
-import de.tum.bgu.msm.models.carOwnership.CreateCarOwnershipModelMuc;
-import de.tum.bgu.msm.models.carOwnership.UpdateCarOwnershipModelMuc;
+import de.tum.bgu.msm.models.*;
 import de.tum.bgu.msm.models.demography.birth.BirthModelImpl;
 import de.tum.bgu.msm.models.demography.birth.DefaultBirthStrategy;
 import de.tum.bgu.msm.models.demography.birthday.BirthdayModel;
@@ -29,6 +23,7 @@ import de.tum.bgu.msm.models.demography.driversLicense.DefaultDriversLicenseStra
 import de.tum.bgu.msm.models.demography.driversLicense.DriversLicenseModel;
 import de.tum.bgu.msm.models.demography.driversLicense.DriversLicenseModelImpl;
 import de.tum.bgu.msm.models.demography.education.EducationModel;
+import de.tum.bgu.msm.models.demography.education.EducationModelImpl;
 import de.tum.bgu.msm.models.demography.employment.EmploymentModel;
 import de.tum.bgu.msm.models.demography.employment.EmploymentModelImpl;
 import de.tum.bgu.msm.models.demography.leaveParentalHousehold.DefaultLeaveParentalHouseholdStrategy;
@@ -38,7 +33,8 @@ import de.tum.bgu.msm.models.demography.marriage.DefaultMarriageStrategy;
 import de.tum.bgu.msm.models.demography.marriage.MarriageModel;
 import de.tum.bgu.msm.models.jobmography.JobMarketUpdate;
 import de.tum.bgu.msm.models.jobmography.JobMarketUpdateImpl;
-import de.tum.bgu.msm.models.realEstate.construction.*;
+import de.tum.bgu.msm.models.realEstate.construction.DefaultConstructionDemandStrategy;
+import de.tum.bgu.msm.models.realEstate.construction.DefaultConstructionLocationStrategy;
 import de.tum.bgu.msm.models.realEstate.demolition.DefaultDemolitionStrategy;
 import de.tum.bgu.msm.models.realEstate.demolition.DemolitionModel;
 import de.tum.bgu.msm.models.realEstate.demolition.DemolitionModelImpl;
@@ -48,21 +44,23 @@ import de.tum.bgu.msm.models.realEstate.pricing.PricingModelImpl;
 import de.tum.bgu.msm.models.realEstate.renovation.DefaultRenovationStrategy;
 import de.tum.bgu.msm.models.realEstate.renovation.RenovationModel;
 import de.tum.bgu.msm.models.realEstate.renovation.RenovationModelImpl;
-import de.tum.bgu.msm.models.relocation.DwellingUtilityStrategyImpl;
-import de.tum.bgu.msm.models.relocation.HousingStrategyMuc;
-import de.tum.bgu.msm.models.relocation.InOutMigrationMuc;
-import de.tum.bgu.msm.models.relocation.RegionUtilityStrategyMucImpl;
-import de.tum.bgu.msm.models.relocation.migration.InOutMigration;
+import de.tum.bgu.msm.models.relocation.migration.InOutMigrationImpl;
 import de.tum.bgu.msm.models.relocation.moves.DefaultDwellingProbabilityStrategy;
 import de.tum.bgu.msm.models.relocation.moves.DefaultMovesStrategy;
 import de.tum.bgu.msm.models.relocation.moves.MovesModelImpl;
+import de.tum.bgu.msm.models.relocation.moves.RegionProbabilityStrategyImpl;
 import de.tum.bgu.msm.models.transportModel.TransportModel;
 import de.tum.bgu.msm.properties.Properties;
+import org.apache.log4j.Logger;
 import org.matsim.core.config.Config;
 
-public class ModelBuilder {
+public class ModelBuilderMstm {
 
-    public static ModelContainer getModelContainerForMuc(DataContainerWithSchools dataContainer, Properties properties, Config config) {
+    private final static Logger logger = Logger.getLogger(ModelBuilderMstm.class);
+
+    private ModelBuilderMstm() {}
+
+    public static ModelContainer getModelContainerForMstm(DataContainer dataContainer, Properties properties, Config config) {
 
         PersonFactory ppFactory = dataContainer.getHouseholdDataManager().getPersonFactory();
         HouseholdFactory hhFactory = dataContainer.getHouseholdDataManager().getHouseholdFactory();
@@ -74,32 +72,33 @@ public class ModelBuilder {
 
         DeathModel deathModel = new DeathModelImpl(dataContainer, properties, new DefaultDeathStrategy());
 
-        MovesModelImpl movesModel = new MovesModelImpl(
-                dataContainer, properties,
-                new DefaultMovesStrategy(),
-                new HousingStrategyMuc(dataContainer,
-                        properties,
-                        dataContainer.getTravelTimes(), new DefaultDwellingProbabilityStrategy(),
-                        new DwellingUtilityStrategyImpl(), new RegionUtilityStrategyMucImpl(), new RegionProbabilityStrategyImpl()));
+        final HousingStrategyMstm housingStrategy = new HousingStrategyMstm(
+                properties,
+                dataContainer,
+                dataContainer.getTravelTimes(),
+                new DefaultDwellingProbabilityStrategy(),
+                new DwellingUtilityStrategyMstm(),
+                new RegionUilityStrategyMstm(), new RegionProbabilityStrategyImpl());
 
-        CreateCarOwnershipModel carOwnershipModel = new CreateCarOwnershipModelMuc(dataContainer);
+        MovesModelImpl movesModel = new MovesModelImpl(dataContainer, properties, new DefaultMovesStrategy(), housingStrategy);
+
 
         DivorceModel divorceModel = new DivorceModelImpl(
-                dataContainer, movesModel, carOwnershipModel, hhFactory,
+                dataContainer, movesModel, null, hhFactory,
                 properties, new DefaultDivorceStrategy());
 
         DriversLicenseModel driversLicenseModel = new DriversLicenseModelImpl(dataContainer, properties, new DefaultDriversLicenseStrategy());
 
-        EducationModel educationModel = new EducationModelMuc(dataContainer, properties);
+        EducationModel educationModel = new EducationModelImpl(dataContainer, properties);
 
         EmploymentModel employmentModel = new EmploymentModelImpl(dataContainer, properties);
 
         LeaveParentHhModel leaveParentsModel = new LeaveParentHhModelImpl(dataContainer, movesModel,
-                carOwnershipModel, hhFactory, properties, new DefaultLeaveParentalHouseholdStrategy());
+                null, hhFactory, properties, new DefaultLeaveParentalHouseholdStrategy());
 
         JobMarketUpdate jobMarketUpdateModel = new JobMarketUpdateImpl(dataContainer, properties);
 
-        ConstructionModel construction = new ConstructionModelImpl(dataContainer, ddFactory,
+        ConstructionModelMstm construction = new ConstructionModelMstm(dataContainer, ddFactory,
                 properties, new DefaultConstructionLocationStrategy(), new DefaultConstructionDemandStrategy());
 
 
@@ -107,35 +106,34 @@ public class ModelBuilder {
 
         RenovationModel renovation = new RenovationModelImpl(dataContainer, properties, new DefaultRenovationStrategy());
 
-        ConstructionOverwrite constructionOverwrite = new ConstructionOverwriteImpl(dataContainer, ddFactory, properties);
+        ConstructionOverwriteMstm constructionOverwrite = new ConstructionOverwriteMstm(dataContainer, ddFactory, properties);
 
-        InOutMigration inOutMigration = new InOutMigrationMuc(dataContainer, employmentModel, movesModel,
-                carOwnershipModel, driversLicenseModel, properties);
+        InOutMigrationImpl inOutMigration = new InOutMigrationImpl(dataContainer, employmentModel, movesModel,
+                null, driversLicenseModel, properties);
 
         DemolitionModel demolition = new DemolitionModelImpl(dataContainer, movesModel,
                 inOutMigration, properties, new DefaultDemolitionStrategy());
 
-        MarriageModel marriageModel = new MarriageModelMuc(dataContainer, movesModel, inOutMigration,
-                carOwnershipModel, hhFactory, properties, new DefaultMarriageStrategy());
+        MarriageModel marriageModel = new MarriageModelMstm(dataContainer, movesModel, inOutMigration,
+                null, hhFactory, properties, new DefaultMarriageStrategy());
 
 
         TransportModel transportModel;
         switch (properties.transportModel.transportModelIdentifier) {
             case MITO_MATSIM:
-                transportModel = new MitoTransportModel(properties.main.baseDirectory, dataContainer, properties, config, new MitoDataConverterMuc());
-                break;
+                logger.warn("Mito not implemented for Mstm. Defaulting to simple matsim transport model");
             case MATSIM:
+                final SimpleMatsimScenarioAssembler scenarioAssembler = new SimpleMatsimScenarioAssembler(dataContainer, properties);
                 transportModel = new MatsimTransportModel(dataContainer, config, properties,
-                		//(MatsimAccessibility) dataContainer.getAccessibility());
-                        null, ZoneConnectorManager.ZoneConnectorMethod.WEIGHTED_BY_POPULATION);
+                        null,
+                        ZoneConnectorManager.ZoneConnectorMethod.RANDOM, scenarioAssembler);
                 break;
             case NONE:
             default:
                 transportModel = null;
-
         }
 
-        final ModelContainer modelContainer = new ModelContainer(
+        return new ModelContainer(
                 birthModel, birthdayModel,
                 deathModel, marriageModel,
                 divorceModel, driversLicenseModel,
@@ -143,9 +141,5 @@ public class ModelBuilder {
                 leaveParentsModel, jobMarketUpdateModel,
                 construction, demolition, pricing, renovation,
                 constructionOverwrite, inOutMigration, movesModel, transportModel);
-
-        modelContainer.registerModelUpdateListener(new UpdateCarOwnershipModelMuc(dataContainer, properties));
-
-        return modelContainer;
     }
 }
