+++ useCases/munich/src/main/java/de/tum/bgu/msm/ModelBuilderMuc.java
@@ -1,13 +1,22 @@
-package run;
+package de.tum.bgu.msm;
 
-import de.tum.bgu.msm.container.DataContainer;
 import de.tum.bgu.msm.container.ModelContainer;
-import de.tum.bgu.msm.data.accessibility.MatsimAccessibility;
+import de.tum.bgu.msm.matsim.MatsimScenarioAssembler;
+import de.tum.bgu.msm.matsim.SimpleMatsimScenarioAssembler;
+import de.tum.bgu.msm.mito.MitoMatsimScenarioAssembler;
+import de.tum.bgu.msm.models.relocation.moves.RegionProbabilityStrategyImpl;
+import de.tum.bgu.msm.schools.DataContainerWithSchools;
 import de.tum.bgu.msm.data.dwelling.DwellingFactory;
 import de.tum.bgu.msm.data.household.HouseholdFactory;
+import de.tum.bgu.msm.data.mito.MitoDataConverterMuc;
 import de.tum.bgu.msm.data.person.PersonFactory;
 import de.tum.bgu.msm.matsim.MatsimTransportModel;
 import de.tum.bgu.msm.matsim.ZoneConnectorManager;
+import de.tum.bgu.msm.models.EducationModelMuc;
+import de.tum.bgu.msm.models.MarriageModelMuc;
+import de.tum.bgu.msm.models.autoOwnership.CreateCarOwnershipModel;
+import de.tum.bgu.msm.models.carOwnership.CreateCarOwnershipModelMuc;
+import de.tum.bgu.msm.models.carOwnership.UpdateCarOwnershipModelMuc;
 import de.tum.bgu.msm.models.demography.birth.BirthModelImpl;
 import de.tum.bgu.msm.models.demography.birth.DefaultBirthStrategy;
 import de.tum.bgu.msm.models.demography.birthday.BirthdayModel;
@@ -29,7 +38,6 @@ import de.tum.bgu.msm.models.demography.leaveParentalHousehold.LeaveParentHhMode
 import de.tum.bgu.msm.models.demography.leaveParentalHousehold.LeaveParentHhModelImpl;
 import de.tum.bgu.msm.models.demography.marriage.DefaultMarriageStrategy;
 import de.tum.bgu.msm.models.demography.marriage.MarriageModel;
-import de.tum.bgu.msm.models.demography.marriage.MarriageModelImpl;
 import de.tum.bgu.msm.models.jobmography.JobMarketUpdate;
 import de.tum.bgu.msm.models.jobmography.JobMarketUpdateImpl;
 import de.tum.bgu.msm.models.realEstate.construction.*;
@@ -42,20 +50,21 @@ import de.tum.bgu.msm.models.realEstate.pricing.PricingModelImpl;
 import de.tum.bgu.msm.models.realEstate.renovation.DefaultRenovationStrategy;
 import de.tum.bgu.msm.models.realEstate.renovation.RenovationModel;
 import de.tum.bgu.msm.models.realEstate.renovation.RenovationModelImpl;
-import de.tum.bgu.msm.models.relocation.migration.InOutMigrationImpl;
-import de.tum.bgu.msm.models.relocation.moves.*;
+import de.tum.bgu.msm.models.relocation.DwellingUtilityStrategyImpl;
+import de.tum.bgu.msm.models.relocation.HousingStrategyMuc;
+import de.tum.bgu.msm.models.relocation.InOutMigrationMuc;
+import de.tum.bgu.msm.models.relocation.RegionUtilityStrategyMucImpl;
+import de.tum.bgu.msm.models.relocation.migration.InOutMigration;
+import de.tum.bgu.msm.models.relocation.moves.DefaultDwellingProbabilityStrategy;
+import de.tum.bgu.msm.models.relocation.moves.DefaultMovesStrategy;
+import de.tum.bgu.msm.models.relocation.moves.MovesModelImpl;
 import de.tum.bgu.msm.models.transportModel.TransportModel;
 import de.tum.bgu.msm.properties.Properties;
-import org.apache.log4j.Logger;
 import org.matsim.core.config.Config;
 
-public class ModelBuilder {
+public class ModelBuilderMuc {
 
-    private final static Logger logger = Logger.getLogger(ModelBuilder.class);
-
-    private ModelBuilder() {}
-
-    public static ModelContainer getModelContainerForMstm(DataContainer dataContainer, Properties properties, Config config) {
+    public static ModelContainer getModelContainerForMuc(DataContainerWithSchools dataContainer, Properties properties, Config config) {
 
         PersonFactory ppFactory = dataContainer.getHouseholdDataManager().getPersonFactory();
         HouseholdFactory hhFactory = dataContainer.getHouseholdDataManager().getHouseholdFactory();
@@ -68,22 +77,27 @@ public class ModelBuilder {
         DeathModel deathModel = new DeathModelImpl(dataContainer, properties, new DefaultDeathStrategy());
 
         MovesModelImpl movesModel = new MovesModelImpl(
-                dataContainer, properties, new DefaultMovesStrategy(), new CarOnlyHousingStrategyImpl(dataContainer, properties, dataContainer.getTravelTimes(), new DwellingUtilityStrategyImpl(),
-                new DefaultDwellingProbabilityStrategy(), new RegionUtilityStrategyImpl(), new RegionProbabilityStrategyImpl()));
+                dataContainer, properties,
+                new DefaultMovesStrategy(),
+                new HousingStrategyMuc(dataContainer,
+                        properties,
+                        dataContainer.getTravelTimes(), new DefaultDwellingProbabilityStrategy(),
+                        new DwellingUtilityStrategyImpl(), new RegionUtilityStrategyMucImpl(), new RegionProbabilityStrategyImpl()));
 
+        CreateCarOwnershipModel carOwnershipModel = new CreateCarOwnershipModelMuc(dataContainer);
 
         DivorceModel divorceModel = new DivorceModelImpl(
-                dataContainer, movesModel, null, hhFactory,
+                dataContainer, movesModel, carOwnershipModel, hhFactory,
                 properties, new DefaultDivorceStrategy());
 
         DriversLicenseModel driversLicenseModel = new DriversLicenseModelImpl(dataContainer, properties, new DefaultDriversLicenseStrategy());
 
-        EducationModel educationModel = null;
+        EducationModel educationModel = new EducationModelMuc(dataContainer, properties);
 
         EmploymentModel employmentModel = new EmploymentModelImpl(dataContainer, properties);
 
         LeaveParentHhModel leaveParentsModel = new LeaveParentHhModelImpl(dataContainer, movesModel,
-                null, hhFactory, properties, new DefaultLeaveParentalHouseholdStrategy());
+                carOwnershipModel, hhFactory, properties, new DefaultLeaveParentalHouseholdStrategy());
 
         JobMarketUpdate jobMarketUpdateModel = new JobMarketUpdateImpl(dataContainer, properties);
 
@@ -97,31 +111,35 @@ public class ModelBuilder {
 
         ConstructionOverwrite constructionOverwrite = new ConstructionOverwriteImpl(dataContainer, ddFactory, properties);
 
-        InOutMigrationImpl inOutMigration = new InOutMigrationImpl(dataContainer, employmentModel, movesModel,
-                null, driversLicenseModel, properties);
+        InOutMigration inOutMigration = new InOutMigrationMuc(dataContainer, employmentModel, movesModel,
+                carOwnershipModel, driversLicenseModel, properties);
 
         DemolitionModel demolition = new DemolitionModelImpl(dataContainer, movesModel,
                 inOutMigration, properties, new DefaultDemolitionStrategy());
 
-        MarriageModel marriageModel = new MarriageModelImpl(dataContainer, movesModel, inOutMigration,
-                null, hhFactory, properties, new DefaultMarriageStrategy());
+        MarriageModel marriageModel = new MarriageModelMuc(dataContainer, movesModel, inOutMigration,
+                carOwnershipModel, hhFactory, properties, new DefaultMarriageStrategy());
 
 
         TransportModel transportModel;
+        MatsimScenarioAssembler scenarioAssembler;
         switch (properties.transportModel.transportModelIdentifier) {
+            case MITO_MATSIM:
+                scenarioAssembler = new MitoMatsimScenarioAssembler(dataContainer, properties, new MitoDataConverterMuc());
+                transportModel = new MatsimTransportModel(dataContainer, config, properties, null,
+                        ZoneConnectorManager.ZoneConnectorMethod.WEIGHTED_BY_POPULATION, scenarioAssembler);
+                break;
             case MATSIM:
-                transportModel = new MatsimTransportModel(dataContainer, config, properties,
-                		(MatsimAccessibility) dataContainer.getAccessibility(),
-                        ZoneConnectorManager.ZoneConnectorMethod.RANDOM);
+                scenarioAssembler = new SimpleMatsimScenarioAssembler(dataContainer, properties);
+                transportModel = new MatsimTransportModel(dataContainer, config, properties, null,
+                        ZoneConnectorManager.ZoneConnectorMethod.WEIGHTED_BY_POPULATION, scenarioAssembler);
                 break;
             case NONE:
-            case MITO_MATSIM:
-                logger.warn("Mito not implemented for mstm. setting transport model to \"NONE\"");
             default:
                 transportModel = null;
-
         }
-        return new ModelContainer(
+
+        final ModelContainer modelContainer = new ModelContainer(
                 birthModel, birthdayModel,
                 deathModel, marriageModel,
                 divorceModel, driversLicenseModel,
@@ -129,5 +147,9 @@ public class ModelBuilder {
                 leaveParentsModel, jobMarketUpdateModel,
                 construction, demolition, pricing, renovation,
                 constructionOverwrite, inOutMigration, movesModel, transportModel);
+
+        modelContainer.registerModelUpdateListener(new UpdateCarOwnershipModelMuc(dataContainer, properties));
+
+        return modelContainer;
     }
 }
