+++ useCases/munich/src/test/java/de/tum/bgu/msm/transportModel/matsim/SiloMucTest.java
@@ -7,7 +7,7 @@ import org.apache.log4j.Logger;
 import org.junit.Test;
 
 import de.tum.bgu.msm.DataBuilder;
-import de.tum.bgu.msm.ModelBuilder;
+import de.tum.bgu.msm.ModelBuilderMuc;
 import de.tum.bgu.msm.SiloModel;
 import de.tum.bgu.msm.container.ModelContainer;
 import de.tum.bgu.msm.io.output.DefaultResultsMonitor;
@@ -38,7 +38,7 @@ public class SiloMucTest {
         DataContainerWithSchools dataContainer = DataBuilder.getModelDataForMuc(properties, null);
         DataBuilder.read(properties, dataContainer);
 
-        ModelContainer modelContainer = ModelBuilder.getModelContainerForMuc(dataContainer, properties, null);
+        ModelContainer modelContainer = ModelBuilderMuc.getModelContainerForMuc(dataContainer, properties, null);
 
         ResultsMonitor resultsMonitor = new DefaultResultsMonitor(dataContainer, properties);
         SiloModel siloModel = new SiloModel(properties, dataContainer, modelContainer, resultsMonitor);
